+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - added new "large" popup template 
 - added a button to empty a layer without deleting it
 - added a button to clone a datalayer
+- added dataUrl and dataFormat on map creation page
 
 ## 0.5.x
 - datalayers are now sent to backend as geojson